requirements.txt
appnope==0.1.4
asttokens==3.0.1
boto3==1.42.62
botocore==1.42.62
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
ipykernel==7.2.0
ipython==8.38.0
jedi==0.19.2
jmespath==1.1.0
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.2.6
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
Pygments==2.19.2
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
s3transfer==0.16.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
